Revert "Merge develop into fix/adq-patch-rte-small-changes"
This reverts commit 441d613ccc7f6c5582c2ce9dfba306711f17e75b, reversing
changes made to f93dc320701bcb49b1b7b31316301cd213a91a8b.

+++ src/solver/optimisation/opt_alloc_probleme_a_optimiser.cpp
@@ -125,7 +125,7 @@ static void optimisationAllocateProblem(PROBLEME_HEBDO* ProblemeHebdo, const int
     logs.info() << " Expected Number of Non-zero terms in Problem Matrix : " << NbTermes;
     logs.info();
 
-    if ((uint)NbTermes > (std::numeric_limits<std::size_t>::max() / 8) - 1)
+    if (NbTermes > (std::numeric_limits<std::size_t>::max() / 8) - 1)
     {
         logs.fatal() << "Optimisation problem too large to be allocated.";
         AntaresSolverEmergencyShutdown();
